Don't lose documents from the spatial index after restart. GC-1
GeoCouch didn't write the current tree height into the header,
hence it couldn't be restored after a restart. This means that
after a restart of GeoCouch a tree was created from scratch
containing only the documents that were inserted after the
restart.

A backwards incompatible change to the on disk format was
needed, therefore all spatial indexes need to be rebuild
after upgrading to this version.

A Futon test is included.

This commit fixes GC-1.