Store FD in spatial record as well.
Only the spatial index group contains the information about
the file descriptor that points to the spatial index file.
This commit adds the information to every index. This way
you don't need to pass on the whole group, but only the index
if you want to read from/write to the file.

Change-Id: I5a575e11731426d840751e765096920ef7933882
Reviewed-on: http://review.couchbase.org/8329
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>