Add skip and limit to spatial merger
The results returned from GeoCouch are not sorted. But
we make sure that always the same results are returned,
even when used with skip and limit. This means that
two subsequent skip request with the correct offset
lead to the same as one big one.

Example:
Request 1: skip1=5, limit1=10
Request 2: skip2=skip1+limit1=15, limit=7
Request 3: skip3=5, limit=17

Request 1 + Request 2 == Request 3

Change-Id: If8814f1515e8e9a23920cffb266cb7bf8e280de5
Reviewed-on: http://review.couchbase.org/11167
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>