Remove source
In the newvtree branch the full source was removed to start
from scratch. Since then the master branch has move on. This
commit does restore that state on current HEAD and then all
changes from the newvtree branch are applied. This way we end
up in a source tree which is exactly the one of the newvtree
branch, but is based on current HEAD. This way a merge commit
is easily possible.

In order to get to the same state as the newvtree branch
originally was, we need to add back the original Makefile that
existed before the move to CMake.

The original commit message in the newvtree branch for this
change was:

    Remove source

    Remove everything except for the Makefile and rebar. The
    new vtree is implemented from scratch.

Change-Id: I2f6d08958c903358498a42883488d49a0b88b274
Reviewed-on: http://review.couchbase.org/15700
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>

Conflicts:
	README.md
	rebar.config
	share/www/script/test/spatial_design_docs.js
	src/geocouch/couch_httpd_spatial.erl
	src/geocouch/couch_httpd_spatial_merger.erl
	src/geocouch/couch_spatial.erl
	src/geocouch/couch_spatial.hrl
	src/geocouch/couch_spatial_compactor.erl
	src/geocouch/couch_spatial_group.erl
	src/geocouch/couch_spatial_merger.erl
	src/geocouch/couch_spatial_updater.erl
	src/geocouch/geocouch_duplicates.erl
	test/100-updater.t
	test/200-compact.t