MB-6860: Don't open group for cleanup
In the old code the spatial group was opened to get the
signature in order to cleanup the files. This meant that
a spatial group was created although there were no spatial
views defined in the design document.

The new code doesn't create new spatial groups if there
aren't eny defined in the design document.

Change-Id: I54f6a3cdfe524c8eeacb88d4bd0d9a87d59985d4
Reviewed-on: http://review.couchbase.org/21591
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>