MB-6860: Fix spatial index file descriptors leaks
On the spatial indexes the file descriptors leaked, when a design
document was updated. The issue (and fix) is similar to the
Apache CouchDB issue COUCHDB-1309 [1].

In Couchbase the design document only lives in the master database,
hence the ddoc_updated event is only triggered for that database,
but we need to clean up the spatial indexes for all vBuckets.

The solution is to save not only a reference of the spatial view
group but also a reference to the database that contains the
design document. This way a cleanup on all vBuckets can be done.

[1] https://issues.apache.org/jira/browse/COUCHDB-1309

Change-Id: Ia9fc277f163a7adbce8a47b0a4212d6ec606858c
Reviewed-on: http://review.couchbase.org/21592
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>