MB-7084: Pass on limit parameter in spatial merger
The limit parameter is now passed on to the other nodes when
the spatial view merger is used. Note that the limit is set
to the original limit plus the original skip. The reason is
that you never know which node needs to skip the results to
which amount (but the maximum number of documents needed is
definitely <= limit + skip).

Change-Id: I7fa9884ef673bbe88c034353ca418ef6843702ed
Reviewed-on: http://review.couchbase.org/22228
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>