Change the on-disk encoding
The encoding is now more similar to the one of the mapreduce views in
order to increase the code re-use in couchstore. Currently the geometry
isn't stored (it will be added later on).

The key is the emitted multidimensional bounding box (MBB) and the
document ID. The value is the partition ID and the emitted value. In
the future the geometry will also be stored as part of the value.

Change-Id: I158b8c35b2c161a46654495c65a95cfaeff2ad2a
Reviewed-on: http://review.couchbase.org/40140
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>