Build geocouch in a special way for the unit tests
The unit tests also test non-exported functions, hence the source needs
to be compiled with the `+export_all` flag. As we don't want to have
everything exported in the normal build, a special build is needed.

With this change it's possible to compile geocouch again with
`GEOCOUCH_BUILD_FOR_UNIT_TESTS=1` set. In order to have everything
work, you need to perform the usual compilation steps frist, which
will create a `build` directory. You then create a subdirectory
(e.g. `geocouch-for-tests`) which will contain the special build.

From within that directory simply run:

    cmake -G "Unix Makefiles" -D BUILD_FOR_UNIT_TESTS=1  ../../geocouch

To run the unit tests, simply run a `make`, followed by a `make test`.

Change-Id: I0d932bd80dd31f355e48cfa15d55fc8bfa3f6b95
Reviewed-on: http://review.couchbase.org/41991
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Volker Mische <volker.mische@gmail.com>