MB-12251: Make inserts into vtree more robust
When inserting several nodes at the same time, the code contained a
special case when the size of the nodes to be inserted was larger than
the maximum threshold, but still less than two times the threshold.

The idea for this special case was that the normal split algorithm could
be used. This assumption was wrong, as it could well be that no valid
split candidate can be found. This resulted in the fatal error
"No split candidates found. Increase the chunk threshold.".

Hence that special case got removed.

This commits also adds a test that would fail with the old logic. Some
other tests used that spatial case also needed to adapted.

Change-Id: Ia4f9e416d5b7b7f27937c7fc9d2a967352d1e6c9
Reviewed-on: http://review.couchbase.org/42896
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>