Make split candidate creation non-fatal
There are cases when there is no split cases, due to the maximum chunk
threshold. It can happen when a single node is bigger than that threshold.
Instead of having a fatal error, create a node that is bigger than that
maximum threshold. This is a trade-off to make. It's better to have a
running system with potentially nodes that are bigger than defined, rather
than a hard crash where no further inserts are possible.

Change-Id: Ic0e4149a5ff967c4e0378532fefdec34b8cf5e05
Reviewed-on: http://review.couchbase.org/42903
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>