MB-12819: Change signature calculation of spatial views
The spatial views now also take the name of the spatial view function
into account (just like mapreduce views). When you change the name of
the function and not the function itself, the view signature will still
change and the view will be re-created.

Re-sorting of the view functions within the JSON object won't change
the function signature.

Change-Id: Ifbe16f9ef2d40eebc6c583aa281e0df6de9e2ad0
Reviewed-on: http://review.couchbase.org/43838
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>