CBIDXT-253: Add wkb library
Add the wkb library in order to encode geometries as Well-Known Binary (WKB).

The library was originally a WKB writer only and created by Cloudant. It's
licensed under the Apache License 2.0.

I extended it with a WKB reader.

This commit includes the full library, but without the build system, as it
originally uses rebar, but Couchbase uses CMake.

Change-Id: I04096132ed7d559bf7dd0ec6d5b137c804e249ca
Reviewed-on: http://review.couchbase.org/45340
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>