MB-12251: Use original MBB for dimensionality comparison
The key of the root node can be be unset. Hence use the original MBB
for the comparison between the query and the index dimensionality.
The original MBB is always set and has the same dimensionality the
index has.

This is a port of 0736029ab47f7dbfa873e16deb4a72079dbe57f0 to the
Apache CouchDB compatible version for GeoCouch.

Change-Id: Idf8043eba608e895d72bf7e541d8892582bb313a